+++ tests/system/example.spec.ts
@@ -1,6 +1,6 @@
 import { spawn } from "node:child_process";
 import { io } from "socket.io-client";
-import { z } from "zod/v4";
+import { z } from "zod";
 import { waitFor } from "../helpers";
 
 describe("System test on Example", async () => {
@@ -40,7 +40,7 @@ describe("System test on Example", async () => {
       await waitFor(() => onError.mock.calls.length > 0);
       expect(onError).toHaveBeenLastCalledWith(
         "InputValidationError",
-        "[0]: Expected function, received undefined",
+        "[0]: Invalid input: expected function, received undefined",
       );
     });
   });